Move `meilisearch_types::Network` to its own module
